feat: Close Delivery request cycles for both "D" and "M"
now delegate can request to deliver a package by sending notification to the e-merchant, and when the e-merchant accept his offer a notification will be sent back to the delegate as an acknowledge.
also e-merchants can request to deliver a certian package with a certian route on the platform by sending a notification to that delegate, and when the delegate accept his offer a notification will be sent back to the merchant as an acknowledge.
packages status will change after each request or acceptance.
the last status right now is reserved.