feat: close all delivery cycle from request to end consumer
now delegate can pick the order via button and it sends a notification to the merchant as an acknowldge for picking, when the delegate deliver the package end-consumer can say to merchant to mark the package as delivered via button and it also sends a notification to the delegate as ana acknowldge to delegate for delivering!!.
phone calls are working now.
only new and negotiable packages appeares in orders list and when merchant want to deliver a package via route!